refactor: move common utils to src
